blackify

+++ pypesto/store/hdf5.py
@@ -6,9 +6,7 @@ from typing import Collection
 from numbers import Number, Integral, Real
 
 
-def write_array(f: h5py.Group,
-                path: str,
-                values: Collection) -> None:
+def write_array(f: h5py.Group, path: str, values: Collection) -> None:
     """
     Write array to hdf5
 
@@ -27,15 +25,12 @@ def write_array(f: h5py.Group,
     elif all(isinstance(x, Real) for x in values):
         write_float_array(f, path, values)
     elif all(isinstance(x, str) for x in values):
-        write_string_array(f, path,
-                           values)
+        write_string_array(f, path, values)
     else:
         f[path] = values
 
 
-def write_string_array(f: h5py.Group,
-                       path: str,
-                       strings: Collection) -> None:
+def write_string_array(f: h5py.Group, path: str, strings: Collection) -> None:
     """
     Write string array to hdf5
 
@@ -50,13 +45,12 @@ def write_string_array(f: h5py.Group,
     """
     dt = h5py.special_dtype(vlen=str)
     dset = f.create_dataset(path, (len(strings),), dtype=dt)
-    dset[:] = [s.encode('utf8') for s in strings]
+    dset[:] = [s.encode("utf8") for s in strings]
 
 
-def write_float_array(f: h5py.Group,
-                      path: str,
-                      values: Collection[Number],
-                      dtype='f8') -> None:
+def write_float_array(
+    f: h5py.Group, path: str, values: Collection[Number], dtype="f8"
+) -> None:
     """
     Write float array to hdf5
 
@@ -78,10 +72,9 @@ def write_float_array(f: h5py.Group,
     dset[:] = values
 
 
-def write_int_array(f: h5py.Group,
-                    path: str,
-                    values: Collection[int],
-                    dtype='<i4'):
+def write_int_array(
+    f: h5py.Group, path: str, values: Collection[int], dtype="<i4"
+):
     """
     Write integer array to hdf5
